feat(types): complete @bizcode/types shared package (#155)
Centralize platform modules/plans catalogs, logistics and IAM DTOs in packages/types.
Update web/server/api-client to consume shared types and reduce server-to-web type imports.
