+++ apps/server/middleware/requireModule.ts
@@ -1,5 +1,5 @@
 import type { NextFunction, Response } from 'express'
-import type { ModuleKey } from '../../web/src/lib/modules'
+import type { ModuleKey } from '@bizcode/types'
 import type { AuthenticatedRequest } from '../auth'
 import { modulesInclude } from '../services/TenantConfigService'
