+++ apps/server/middleware/resolvePortalTenant.ts
@@ -1,7 +1,7 @@
 import type { NextFunction, Response } from 'express'
 import type { PrismaClient } from '@prisma/client'
-import type { ModuleKey } from '../../web/src/lib/modules'
-import { NEW_TENANT_MODULES } from '../../web/src/lib/modules/tenantDefaults'
+import type { ModuleKey } from '@bizcode/types'
+import { NEW_TENANT_MODULES } from '@bizcode/types'
 import { PortalConfigService } from '../services/PortalConfigService'
 import { TenantConfigService, modulesInclude } from '../services/TenantConfigService'
 import type { PortalRequest } from '../portal/portalTypes'
