+++ apps/server/routes/registerArticulosRoutes.ts
@@ -3,7 +3,7 @@ import { requirePermission, type AuthenticatedRequest } from '../auth'
 import { validateBody } from '../middleware/validateBody'
 import { articuloBodySchema, stockAjusteBodySchema } from '../schemas/domain'
 import type { ArticuloInput, StockAjusteInput } from '../createApp.types'
-import { hasPermission } from '../../web/src/lib/rbac'
+import { hasPermission } from '@bizcode/types'
 import { readDbfRecordsFromBuffer } from '../../web/src/lib/migration/readDbfBuffer'
 import { parseCsvWithFixedHeaders, CSV_IMPORT_MAX_ROWS } from '../csvImport'
 import { paginatedListJson, parseListPagination } from '../services/listPagination'
