+++ apps/server/routes/registerRepartosRoutes.ts
@@ -2,7 +2,7 @@ import type { Application, Request, Response } from 'express'
 import { requirePermission } from '../auth'
 import type { AuthenticatedRequest } from '../auth'
 import { validateBody } from '../middleware/validateBody'
-import { hasPermission, type Permission } from '../../web/src/lib/rbac'
+import { hasPermission, type Permission } from '@bizcode/types'
 import { requireModule } from '../middleware/requireModule'
 import { repartoCreateBodySchema, repartoItemPodBodySchema, repartoUbicacionBodySchema } from '../schemas/domain'
 import { GPS_VIEW_ROLES, POD_VIEW_ROLES } from '../services/RepartoService'
