+++ apps/server/services/tenantConfigCache.ts
@@ -1,4 +1,4 @@
-import type { ModuleKey } from '../../web/src/lib/modules'
+import type { ModuleKey } from '@bizcode/types'
 
 type CacheEntry = {
   modules: readonly ModuleKey[]
