+++ apps/web/src/lib/modules/types.ts
@@ -1,30 +1,13 @@
-export const MODULE_PLANS = ['starter', 'pro', 'enterprise'] as const
-
-export type ModulePlan = (typeof MODULE_PLANS)[number]
-
-export const DEPLOYMENT_ENVS = ['dev', 'prod'] as const
-
-export type DeploymentEnv = (typeof DEPLOYMENT_ENVS)[number]
-
-export interface ModuleDef {
-  label: string
-  required: boolean
-  requiredInProd: boolean
-  dependencies: readonly string[]
-  plan: ModulePlan
-  price: number
-}
-
-export type ModuleValidationReason =
-  | 'required_module_missing'
-  | `missing_dependency:${string}`
-
-export interface ModuleValidationError {
-  module: string
-  reason: ModuleValidationReason
-}
-
-export interface ModuleValidationResult {
-  valid: boolean
-  errors: ModuleValidationError[]
-}
+export {
+  DEPLOYMENT_ENVS,
+  MODULE_PLANS,
+  type DeploymentEnv,
+  type ModuleCatalogEntry,
+  type ModuleDef,
+  type ModulePlan,
+  type ModuleValidationError,
+  type ModuleValidationReason,
+  type ModuleValidationResult,
+  type TenantMonthlyPriceEstimate,
+  type TenantPricingAddon,
+} from '@bizcode/types'
