+++ packages/types/src/modules-catalog.ts
@@ -1,0 +1,422 @@
+import type { ModuleDef } from './modules'
+
+export const MODULE_CATALOG = {
+  'core.auth': {
+    label: 'Auth & RBAC',
+    required: true,
+    requiredInProd: true,
+    dependencies: [],
+    plan: 'starter',
+    price: 0,
+  },
+  'core.catalog': {
+    label: 'Catalogo',
+    required: true,
+    requiredInProd: true,
+    dependencies: [],
+    plan: 'starter',
+    price: 0,
+  },
+  'core.clients': {
+    label: 'Clientes',
+    required: true,
+    requiredInProd: true,
+    dependencies: [],
+    plan: 'starter',
+    price: 0,
+  },
+  'core.invoicing': {
+    label: 'Facturacion base',
+    required: true,
+    requiredInProd: true,
+    dependencies: [],
+    plan: 'starter',
+    price: 0,
+  },
+  'billing.arca_cae': {
+    label: 'ARCA CAE',
+    required: false,
+    requiredInProd: true,
+    dependencies: ['core.invoicing'],
+    plan: 'starter',
+    price: 0,
+  },
+  'billing.credit_notes': {
+    label: 'Notas de credito',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['billing.arca_cae'],
+    plan: 'starter',
+    price: 0,
+  },
+  'billing.orders': {
+    label: 'Pedidos',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.catalog', 'core.clients'],
+    plan: 'starter',
+    price: 1500,
+  },
+  'billing.pos': {
+    label: 'Punto de venta',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.invoicing'],
+    plan: 'starter',
+    price: 2000,
+  },
+  'finance.collections': {
+    label: 'Cobranzas',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.invoicing'],
+    plan: 'starter',
+    price: 2000,
+  },
+  'finance.receipts': {
+    label: 'Recibos de pago',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['finance.collections'],
+    plan: 'starter',
+    price: 0,
+  },
+  'finance.ledger': {
+    label: 'Cuenta corriente',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['billing.arca_cae'],
+    plan: 'pro',
+    price: 2500,
+  },
+  'finance.bank_reconcile': {
+    label: 'Conciliacion bancaria',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['finance.ledger'],
+    plan: 'pro',
+    price: 3000,
+  },
+  'finance.retenciones': {
+    label: 'Retenciones/percepciones',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['billing.arca_cae'],
+    plan: 'starter',
+    price: 0,
+  },
+  'fiscal.remito': {
+    label: 'Remito de entrega',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.invoicing'],
+    plan: 'starter',
+    price: 0,
+  },
+  'fiscal.cheques': {
+    label: 'Gestion de cheques',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['finance.collections'],
+    plan: 'starter',
+    price: 0,
+  },
+  'logistics.dispatches': {
+    label: 'Repartos',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.clients'],
+    plan: 'pro',
+    price: 2500,
+  },
+  'logistics.picking': {
+    label: 'Picking',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['logistics.dispatches', 'inventory.stock'],
+    plan: 'pro',
+    price: 1500,
+  },
+  'logistics.pod': {
+    label: 'POD firma',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['logistics.dispatches'],
+    plan: 'pro',
+    price: 1000,
+  },
+  'logistics.gps': {
+    label: 'Tracking GPS',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['logistics.dispatches'],
+    plan: 'enterprise',
+    price: 5000,
+  },
+  'logistics.purchases': {
+    label: 'Ordenes de compra',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.catalog'],
+    plan: 'starter',
+    price: 1500,
+  },
+  'inventory.stock': {
+    label: 'Gestion de stock',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.catalog'],
+    plan: 'starter',
+    price: 1500,
+  },
+  'inventory.count': {
+    label: 'Recuento fisico',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['inventory.stock'],
+    plan: 'pro',
+    price: 1000,
+  },
+  'inventory.fefo': {
+    label: 'FEFO / Lotes',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['inventory.stock'],
+    plan: 'pro',
+    price: 2000,
+  },
+  'inventory.lots': {
+    label: 'Trazabilidad lotes',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['inventory.fefo'],
+    plan: 'enterprise',
+    price: 3000,
+  },
+  'apps.driver': {
+    label: 'App Repartidor',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['logistics.dispatches'],
+    plan: 'pro',
+    price: 4000,
+  },
+  'apps.seller': {
+    label: 'App Vendedor',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['billing.orders'],
+    plan: 'pro',
+    price: 4000,
+  },
+  'comms.notifications': {
+    label: 'Notificaciones in-app',
+    required: false,
+    requiredInProd: false,
+    dependencies: [],
+    plan: 'starter',
+    price: 0,
+  },
+  'comms.whatsapp': {
+    label: 'WhatsApp Twilio',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['comms.notifications'],
+    plan: 'pro',
+    price: 2000,
+  },
+  'comms.chat': {
+    label: 'Chat interno',
+    required: false,
+    requiredInProd: false,
+    dependencies: [],
+    plan: 'starter',
+    price: 0,
+  },
+  'analytics.dashboard': {
+    label: 'Dashboard basico',
+    required: false,
+    requiredInProd: false,
+    dependencies: [],
+    plan: 'starter',
+    price: 0,
+  },
+  'analytics.advanced': {
+    label: 'Analytics avanzado',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['analytics.dashboard'],
+    plan: 'pro',
+    price: 3000,
+  },
+  'admin.audit_log': {
+    label: 'Log de auditoria',
+    required: false,
+    requiredInProd: false,
+    dependencies: [],
+    plan: 'starter',
+    price: 0,
+  },
+  'service.contracts': {
+    label: 'Contratos recurrentes',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.invoicing'],
+    plan: 'pro',
+    price: 2000,
+  },
+  'service.orders': {
+    label: 'Ordenes de trabajo',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.clients'],
+    plan: 'pro',
+    price: 2000,
+  },
+  'service.warranties': {
+    label: 'Garantias',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.invoicing'],
+    plan: 'starter',
+    price: 0,
+  },
+  'pos.cashier': {
+    label: 'Arqueo de caja / turnos',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['billing.pos'],
+    plan: 'starter',
+    price: 1000,
+  },
+  'production.bom': {
+    label: 'Formulas BOM',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['inventory.stock'],
+    plan: 'pro',
+    price: 2500,
+  },
+  'production.orders': {
+    label: 'Ordenes de produccion',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['production.bom'],
+    plan: 'pro',
+    price: 2500,
+  },
+  'clients.loyalty': {
+    label: 'Programa de fidelizacion',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.clients'],
+    plan: 'pro',
+    price: 2000,
+  },
+  'clients.portal': {
+    label: 'Portal del cliente',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.invoicing', 'finance.ledger'],
+    plan: 'pro',
+    price: 3000,
+  },
+  'catalog.multicurrency': {
+    label: 'Precios USD/EUR + TC',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['core.catalog'],
+    plan: 'starter',
+    price: 1500,
+  },
+  'analytics.ai_demand': {
+    label: 'IA Prediccion demanda',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['analytics.advanced'],
+    plan: 'enterprise',
+    price: 8000,
+  },
+  'analytics.ai_routes': {
+    label: 'IA Rutas TSP',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['logistics.dispatches'],
+    plan: 'enterprise',
+    price: 5000,
+  },
+  'analytics.ai_anomaly': {
+    label: 'IA Anomalias',
+    required: false,
+    requiredInProd: false,
+    dependencies: ['analytics.advanced'],
+    plan: 'enterprise',
+    price: 5000,
+  },
+  'analytics.ai_chatbot': {
+    label: 'IA Chatbot atencion',
+    required: false,
+    requiredInProd: false,
+    dependencies: [],
+    plan: 'enterprise',
+    price: 5000,
+  },
+  'platform.white_label': {
+    label: 'White-label',
+    required: false,
+    requiredInProd: false,
+    dependencies: [],
+    plan: 'enterprise',
+    price: 0,
+  },
+  'platform.webhooks': {
+    label: 'Webhooks salientes',
+    required: false,
+    requiredInProd: false,
+    dependencies: [],
+    plan: 'pro',
+    price: 0,
+  },
+  'platform.api_public': {
+    label: 'API publica (OpenAPI)',
+    required: false,
+    requiredInProd: false,
+    dependencies: [],
+    plan: 'enterprise',
+    price: 0,
+  },
+} as const satisfies Record<string, ModuleDef>
+
+export type ModuleKey = keyof typeof MODULE_CATALOG
+
+export const MODULE_KEYS = Object.keys(MODULE_CATALOG) as ModuleKey[]
+
+export const DEFAULT_MODULES = [
+  'core.auth',
+  'core.catalog',
+  'core.clients',
+  'core.invoicing',
+  'billing.arca_cae',
+  'billing.credit_notes',
+  'finance.collections',
+  'finance.receipts',
+  'finance.ledger',
+  'fiscal.remito',
+  'fiscal.cheques',
+  'finance.retenciones',
+  'comms.notifications',
+  'analytics.dashboard',
+  'admin.audit_log',
+  'service.warranties',
+] as const satisfies readonly ModuleKey[]
+
+export type ModuleCatalogEntry = {
+  key: ModuleKey
+  label: string
+  required: boolean
+  requiredInProd: boolean
+  dependencies: readonly ModuleKey[]
+  plan: ModuleDef['plan']
+  price: number
+  canDeactivate: boolean
+}
